requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
aws-cdk.assets==1.16.3
aws-cdk.aws-cloudwatch==1.16.3
aws-cdk.aws-ec2==1.16.3
aws-cdk.aws-events==1.16.3
aws-cdk.aws-iam==1.16.3
aws-cdk.aws-kms==1.16.3
aws-cdk.aws-lambda==1.16.3
aws-cdk.aws-secretsmanager==1.16.3
aws-cdk.aws-codepipeline==1.16.3
aws-cdk.aws-codepipeline-actions==1.16.3
aws-cdk.aws-logs==1.16.3
aws-cdk.aws-s3==1.16.3
aws-cdk.aws-s3-assets==1.16.3
aws-cdk.aws-sqs==1.16.3
aws-cdk.aws-ssm==1.16.3
aws-cdk.core==1.16.3
aws-cdk.cx-api==1.16.3
aws-cdk.region-info==1.16.3
cattrs==0.9.0
importlib-metadata==0.19
jsii==0.14.3
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
publication==0.0.3
py==1.8.0
pyparsing==2.4.2
pytest==5.0.1
python-dateutil==2.8.0
six==1.12.0
typing-extensions==3.7.4
wcwidth==0.1.7
zipp==0.5.2